# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KPI tizimi API - Kattaqo'rg'on tumani suv yetkazib berish xizmati
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kpi-categories:
    get:
      operationId: listKpiCategories
      tags: [kpi-categories]
      summary: List KPI categories
      responses:
        "200":
          description: List of KPI categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiCategory"
    post:
      operationId: createKpiCategory
      tags: [kpi-categories]
      summary: Create KPI category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiCategory"
  /kpi-categories/{id}:
    put:
      operationId: updateKpiCategory
      tags: [kpi-categories]
      summary: Update KPI category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiCategory"
    delete:
      operationId: deleteKpiCategory
      tags: [kpi-categories]
      summary: Delete KPI category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /kpi-indicators:
    get:
      operationId: listKpiIndicators
      tags: [kpi-indicators]
      summary: List KPI indicators
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of KPI indicators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiIndicator"
    post:
      operationId: createKpiIndicator
      tags: [kpi-indicators]
      summary: Create KPI indicator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiIndicatorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiIndicator"
  /kpi-indicators/{id}:
    put:
      operationId: updateKpiIndicator
      tags: [kpi-indicators]
      summary: Update KPI indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKpiIndicatorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiIndicator"
    delete:
      operationId: deleteKpiIndicator
      tags: [kpi-indicators]
      summary: Delete KPI indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /evaluations:
    get:
      operationId: listEvaluations
      tags: [evaluations]
      summary: List evaluations
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
        - name: period
          in: query
          required: false
          schema:
            type: string
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evaluation"
    post:
      operationId: createEvaluation
      tags: [evaluations]
      summary: Create evaluation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
  /evaluations/{id}:
    get:
      operationId: getEvaluation
      tags: [evaluations]
      summary: Get evaluation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
    put:
      operationId: updateEvaluation
      tags: [evaluations]
      summary: Update evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
    delete:
      operationId: deleteEvaluation
      tags: [evaluations]
      summary: Delete evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    KpiCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        weight:
          type: number
        color:
          type: string
          nullable: true
        indicatorCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - weight
        - indicatorCount
        - createdAt
    CreateKpiCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        weight:
          type: number
        color:
          type: string
          nullable: true
      required:
        - name
        - weight
    KpiIndicator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        targetValue:
          type: number
          nullable: true
        weight:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - categoryId
        - weight
        - createdAt
    CreateKpiIndicatorBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
        unit:
          type: string
          nullable: true
        targetValue:
          type: number
          nullable: true
        weight:
          type: number
      required:
        - name
        - categoryId
        - weight
    Evaluation:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        departmentName:
          type: string
          nullable: true
        indicatorId:
          type: integer
        indicatorName:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        period:
          type: string
        score:
          type: number
        maxScore:
          type: number
        comment:
          type: string
          nullable: true
        evaluatorId:
          type: integer
          nullable: true
        evaluatorName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - indicatorId
        - period
        - score
        - maxScore
        - createdAt
    CreateEvaluationBody:
      type: object
      properties:
        employeeId:
          type: integer
        indicatorId:
          type: integer
        period:
          type: string
        score:
          type: number
        maxScore:
          type: number
        comment:
          type: string
          nullable: true
      required:
        - employeeId
        - indicatorId
        - period
        - score
        - maxScore
...